Refactor trade execution and strategy scores
