+++ live/scheduler_loop.py
@@ -54,10 +54,9 @@ from connectors.mt5_connector import get_account_info
 from utils.trade_journal import record_trade, update_trade, load_history
 from utils.logger import log_trade_action
 from risk_management.breakeven_manager import BreakEvenManager
-from execution.order_manager import execute_fake_order
 from ai_engine.strategy_score_manager import recover_old_scores
+from core.trade_engine import execute_trade
 from monitoring.alert_manager import (
-    alert_trade_opened,
     alert_sl_moved,
     alert_trade_closed,
     alert_daily_guard,
@@ -83,39 +82,6 @@ from strategy_components.session_filter import (
 logger = logging.getLogger("scheduler")
 
 
-def is_valid_sl_tp(symbol: str, order_type: str, price: float, sl: float, tp: float) -> bool:
-    """Validate SL/TP against broker stop levels."""
-    info = mt5.symbol_info(symbol)
-    if not info:
-        return False
-    stop_level = getattr(info, "trade_stops_level", getattr(info, "stops_level", 0)) * info.point
-
-    if order_type == "buy":
-        if sl >= price or tp <= price:
-            return False
-        if (price - sl) < stop_level or (tp - price) < stop_level:
-            return False
-    else:
-        if sl <= price or tp >= price:
-            return False
-        if (sl - price) < stop_level or (price - tp) < stop_level:
-            return False
-    return True
-
-
-def auto_adjust_sl_tp(symbol: str, order_type: str, price: float, sl: float, tp: float) -> tuple[float, float]:
-    """Adjust SL/TP to satisfy minimum stop level requirements."""
-    info = mt5.symbol_info(symbol)
-    if not info:
-        return sl, tp
-    min_dist = getattr(info, "trade_stops_level", getattr(info, "stops_level", 0)) * info.point
-    if order_type == "buy":
-        sl = min(sl, price - min_dist)
-        tp = max(tp, price + min_dist)
-    else:
-        sl = max(sl, price + min_dist)
-        tp = min(tp, price - min_dist)
-    return sl, tp
 
 def get_confidence_threshold(symbol: str, timeframe: str, regime: str, strategy_name: str) -> float:
     """Return confidence threshold for a trade context."""
@@ -209,111 +175,6 @@ def refresh_data(symbol: str, timeframe: str, limit: int = 300) -> None:
     enriched = add_indicators(clean)
     indicator_cache[(symbol, timeframe)] = enriched[symbol][timeframe]
 
-def execute_trade(
-    direction: str, symbol: str, lot: float, sl: float, tp: float, magic_offset: int = 0
-) -> int | None:
-    if not is_market_open(symbol):
-        logger.warning(f"[SKIP] Market is closed for {symbol}, skipping trade.")
-        return None
-    info = mt5.symbol_info(symbol)
-    if not info:
-        logger.error(f"\u26a0\ufe0f Failed to fetch SymbolInfo for {symbol}")
-        return None
-    trade_mode = getattr(info, "trade_mode", None)
-    if trade_mode in (0, 3):
-        logger.warning(f"{symbol} not tradeable now (market closed or disabled)")
-        return None
-
-    price_tick = mt5.symbol_info_tick(symbol)
-    if price_tick is None:
-        logging.warning(f"No price data for {symbol}")
-        return None
-
-    entry_price = price_tick.ask if direction == "buy" else price_tick.bid
-
-    if not is_valid_sl_tp(symbol, direction, entry_price, sl, tp):
-        info = mt5.symbol_info(symbol)
-        stop_level = info.trade_stops_level * info.point if info else 0
-        print(f"[VALIDATION FAILED] {symbol} ({direction})")
-        print(
-            f"Entry={entry_price:.2f}, SL={sl:.2f}, TP={tp:.2f}, MinStop={stop_level:.2f}"
-        )
-        print(
-            f"\u274c Invalid SL/TP for {symbol} - Order rejected due to broker stop limits."
-        )
-        return None
-
-    sl, _, valid = enforce_min_stop_distance(symbol, entry_price, sl, tp, direction)
-    if not valid:
-        return None
-
-    print(f"\U0001F4E4 Executing {direction.upper()} on {symbol} @ {entry_price:.2f}")
-    print(f"    SL: {sl:.2f} | TP: {tp:.2f}")
-    if get_config("LIVE_MODE", "false").lower() != "true":
-        logger.debug(
-            "[SIM MODE] %s %s lot=%.2f price=%.5f sl=%s tp=%s",
-            direction,
-            symbol,
-            lot,
-            entry_price,
-            sl,
-            tp,
-        )
-        execute_fake_order(direction, symbol, lot, entry_price, sl=sl, tp=tp)
-        alert_trade_opened(symbol, "sim", direction, entry_price, sl, tp)
-        return int(time.time())
-
-    deal_type = mt5.ORDER_TYPE_BUY if direction == "buy" else mt5.ORDER_TYPE_SELL
-    symbol_info = mt5.symbol_info(symbol)
-    type_filling = mt5.ORDER_FILLING_IOC  # default fallback
-    # Optional: make it dynamic per-symbol from config
-    override = get_config("FILLING_MODE_OVERRIDES", {})
-    if symbol in override:
-        type_filling = int(override[symbol])
-    elif symbol_info and hasattr(symbol_info, "filling_mode"):
-        if symbol_info.filling_mode in (
-            mt5.ORDER_FILLING_IOC,
-            mt5.ORDER_FILLING_FOK,
-            mt5.ORDER_FILLING_RETURN,
-        ):
-            type_filling = symbol_info.filling_mode
-    request = {
-        "action": mt5.TRADE_ACTION_DEAL,
-        "symbol": symbol,
-        "volume": lot,
-        "type": deal_type,
-        "price": entry_price,
-        "sl": sl,
-        "tp": tp,
-        "deviation": 20,
-        "magic": MAGIC_NUMBER + magic_offset,
-        "comment": "AI Trade",
-        "type_time": mt5.ORDER_TIME_GTC,
-        "type_filling": type_filling,
-    }
-    if tp and tp > 0:
-        request["tp"] = tp
-    info = mt5.symbol_info(symbol)
-    if not info or info.trade_mode != mt5.SYMBOL_TRADE_MODE_FULL:
-        logger.warning(f"{symbol} is not tradeable now (market closed or disabled)")
-        return None
-    logger.info(f"Sending order for {symbol}...")
-    logger.debug("Order details: %s", request)
-    result = mt5.order_send(request)
-    logger.debug(
-        "MT5 retcode=%s comment=%s", getattr(result, "retcode", None), getattr(result, "comment", "")
-    )
-    if result and result.retcode == mt5.TRADE_RETCODE_DONE:
-        logger.info("Trade executed: ticket %s", result.order)
-        alert_trade_opened(symbol, "live", direction, entry_price, sl, tp)
-        return result.order
-    logger.error(
-        "Trade failed: retcode=%s reason=%s response=%s",
-        getattr(result, "retcode", None),
-        getattr(result, "comment", ""),
-        result,
-    )
-    return None
 
 
 def process_symbol_timeframe(symbol: str, timeframe: str, *, force_trade: bool = False) -> None:
